% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_media.R
\name{find_duplicate_media}
\alias{find_duplicate_media}
\title{Find Duplicate Media Calls in Rmd Slide Show (if any) as Aid to Removing Duplicates; Optionally, Save to File in Current Directory}
\usage{
find_duplicate_media(rmd_file = "sample-slides.rmd", save_to_file = FALSE)
}
\arguments{
\item{rmd_file}{An .Rmd path/filename provided as a character string.  Default value = "sample-slides.Rmd"}

\item{save_to_file}{A logical vector,provided as a character string, indicating user request to write a new file to current working directory.  Default value = "FALSE"}
}
\value{
An R object named 'dupliate_media' which is printed in console, or (optionally) written to a new Rmd file in current working directory.
}
\description{
Finds duplicate media calls in Rmd slide deck (if any) and reports them in the console as aid to removing duplicates.  Optionally, will save duplicates to file in current directory.  User supplies argument for Rmd path/filename containing the slide deck, and the (optional) logical argument of TRUE.  Default value = FALSE.    CAUTION - WHEN 'save_to_file' argument = TRUE, A NEW FILE IS SAVED IN THE CURRENT DIRECTORY NAMED 'media-duplicates.rmd'.

Confirmation message of saved file appears in the console when 'save_to_file' argument = TRUE.  Otherwise, list of duplicates (if any) are written to the console.
}
\examples{
rmd_file = system.file("extdata", "sample-slides.rmd", package = "xtractr")
find_duplicate_media(rmd_file)

\dontrun{
   find_duplicate_media("path/to/your/sample-slides.rmd", save_to_file = TRUE)
}

}
