% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media_calls.R
\name{get_media_calls}
\alias{get_media_calls}
\title{Extract Paths and Filenames of All Media Calls in Rmd File}
\usage{
get_media_calls(rmd_file)
}
\arguments{
\item{rmd_file}{Path and filename of the Rmarkdown file to be processed. Rmd file must be present in the current working directory.}
}
\value{
Character vector of path/filename of extracted media calls prints in the console
}
\description{
Prints paths and filenames of all media calls found in Rmd file into the console to be available for copy and paste operations.
}
\examples{
rmd_file = system.file("extdata", "sample-slides.rmd", package = "xtractr")
get_media_calls(rmd_file)
}
