% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmt_gif_loops.R
\name{lmt_gif_loops}
\alias{lmt_gif_loops}
\title{Limit GIF Loops}
\usage{
lmt_gif_loops(input_gif, output_gif, loop_limit = 3)
}
\arguments{
\item{input_gif}{Path/filename of the input GIF.}

\item{output_gif}{Path/filename where the limited-loop GIF will be saved.}

\item{loop_limit}{(optional) The number of loops to limit the GIF to (default is 3).}
}
\value{
None (the function saves the modified GIF to a new output file).
}
\description{
This function takes the path/filename of a GIF, limits the number of loops, and saves the modified GIF to a new file specified by user.
}
\examples{
\dontrun{
# Limit the loops to 2 and save the result to a new file
lmt_gif_loops("gif/red-raincoat-in-crowd.gif", "gif/ltd-red-raincoat-in-crowd.gif", loop_limit = 2)
}

}
