% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_to_bookdown.R
\name{slides_to_bookdown}
\alias{slides_to_bookdown}
\title{Convert Rmd Slide Show to Bookdown Compatible Text for Further Editing}
\usage{
slides_to_bookdown(
  rmd_file = "sample-slides.rmd",
  txt_output_file = "sample-slides-to-bookdown.rmd",
  save_to_file = FALSE
)
}
\arguments{
\item{rmd_file}{An .Rmd path/filename provided as a character string.  Default value = "sample-slides.Rmd"}

\item{txt_output_file}{An Rmd path/filename provided as a character string.  Default value = "sample-slides-to-bookdown.Rmd"}

\item{save_to_file}{A logical vector,provided as a character string, indicating user request to write a new file to current working directory.  Default value = "FALSE"}
}
\value{
An R object named 'modified_text' which is (optionally) written to a new Rmd file, suitable for editing in bookdown.  New file is created in current working directory.
}
\description{
Rmd Slide Show file is transformed to bookdown compatible text with no more than two spaces between words and limited to 80 characters per line.  Output appears in console for further 'copy and paste' operation.  Optionally, the transformed text can be stored in new Rmd file in the current working directory.  User supplies arguments for Rmd path/filename containing the slide deck, and the (optional) output path/filename.  Default values are "sample-slides.Rmd" and "sample-slides-to-bookdown.Rmd".  CAUTION - WHEN 'save_to_file' argument = TRUE, TWO NEW INTERIM FILES ARE SAVED IN THE CURRENT DIRECTORY:  rmd_file.pdf and rmd_file.html.

Confirmation message of saved file appears in the console.
}
\examples{

\dontrun{
   slides_to_bookdown(rmd_file) # prints modified text to console
   slides_to_bookdown("sample-slides.rmd", "sample-slides-to-bookdown.rmd", save_to_file = TRUE)
}

}
