% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_to_bookdown_txt.R
\name{slides_to_bookdown_txt}
\alias{slides_to_bookdown_txt}
\title{Convert Rmd Slide Show to Bookdown Compatible Text for Further Editing}
\usage{
slides_to_bookdown_txt(
  rmd_file = "sample-slides-copy.rmd",
  txt_output_file = "sample-slides-to-bookdown.Rmd"
)
}
\arguments{
\item{rmd_file}{An .Rmd path/filename provided as a character string.  Default value = "sample-slides-copy.Rmd"}

\item{txt_output_file}{An Rmd path/filename provided as a character string.  Default value = "sample-slides-to-bookdown.Rmd"}
}
\value{
An R object named modified_text which is written to a new Rmd file, suitable for editing in bookdown.  New file is created in current working directory.
}
\description{
Rmd Slide Show file is transformed to single line of bookdown compatible text and stored in new Rmd file in the current working directory.  User supplies argument for Rmd path/filename containing the slide deck, and the argument for the output path/filename.  Default values are "sample-slides-copy.Rmd" and "sample-slides-to-bookdown.Rmd".

Confirmation messages appear in the console.
}
\examples{
\dontrun{
   slides_to_bookdown_txt("sample-slides-copy.rmd", "sample-slides-to-bookdown.Rmd")
}

}
