% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_media_calls.R
\name{transform_media_calls}
\alias{transform_media_calls}
\title{Transforms All Media Calls in Rmd File to Rmarkdown Syntax}
\usage{
transform_media_calls(rmd_file)
}
\arguments{
\item{rmd_file}{An .Rmd path/filename provided as a character string.}
}
\value{
Character vector prints in the console
}
\description{
Prints all media calls found in Rmd file into the console as originally found and in Rmarkdown \code{![]()} format to be available for copy and paste operations.
}
\examples{
rmd_file = system.file("extdata", "sample-slides.rmd", package = "xtractr")
transform_media_calls(rmd_file)
}
