% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtract_gifs_optional_save.R
\name{xtract_gifs_optional_save}
\alias{xtract_gifs_optional_save}
\title{Extract GIF Media Calls Found in Rmd File, Transform to Rmarkdown \code{![]()} format and Write to Console; Optionally Write to Disk}
\usage{
xtract_gifs_optional_save(rmd_file, output_file, write_to_file = FALSE)
}
\arguments{
\item{rmd_file}{An .Rmd path/filename provided as a character string.}

\item{output_file}{An .Rmd path/filename provided as a character string. File is saved in the current working directory.}

\item{write_to_file}{A logical value indicating whether to write the result to a file (default = FALSE).}
}
\value{
Character vector object named 'gifs-only'.
}
\description{
All GIF media calls found in the specified Rmd file are transformed into Rmarkdown \code{![]()} syntax. The list is sent to the console as originally found and in the Rmarkdown \code{![]()} format. If \code{write_to_file} is \code{TRUE} (default = FALSE) and 'output_file' argument has a text value, the list is saved to a new file with the specified \code{output_file} name. If \code{write_to_file} is \code{FALSE}, no file is written.
}
\examples{
rmd_file <- system.file("extdata", "sample-slides.rmd", package = "xtractr")
# Only print to console, no file is saved
xtract_gifs_optional_save(rmd_file)
#' # Save to a file with a custom filename "custom-output.rmd"
\dontrun{
xtract_gifs_optional_save(rmd_file, output_file = "custom-output.rmd", write_to_file = TRUE)
}
}
