% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractr-package.R
\docType{package}
\name{xtractr-package}
\alias{xtractr}
\alias{xtractr-package}
\title{xtractr: Helper Functions for Extracting Textual Content from Rmd Files, Including Slide Decks}
\description{
Helper functions to extract textual content from Rmd files, including slides.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danswart/xtractr}
  \item Report bugs at \url{https://github.com/danswart/xtractr/issues}
}

}
\author{
\strong{Maintainer}: Dan Swart \email{danswartcpa@gmail.com} (\href{https://orcid.org/0009-0000-6802-3979}{ORCID})

}
\keyword{internal}
