% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractr-package.R
\docType{package}
\name{xtractr-package}
\alias{xtractr}
\alias{xtractr-package}
\title{xtractr: Helper Functions for Creating Presentation Slides and Bookdown Projects Including Extracting & Embedding Textual Content To and From Rmd Files}
\description{
This package contains helper functions for creating presentation slides and bookdown projects including extracting & embedding textual content to and from Rmd files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danswart/xtractr}
  \item Report bugs at \url{https://github.com/danswart/xtractr/issues}
}

}
\author{
\strong{Maintainer}: Dan Swart \email{danswartcpa@gmail.com} (\href{https://orcid.org/0009-0000-6802-3979}{ORCID})

}
\keyword{internal}
